requirements.txt
a2a-sdk[http-server]
google-generativeai
fastapi
uvicorn
python-dotenv
pydantic
pytest
pytest-asyncio
starlette
